handle all 3 player interaction events

+++ arena-api/src/main/java/io/github/zap/arenaapi/anticheat/classifier/Classifier.java
@@ -2,6 +2,6 @@ package io.github.zap.arenaapi.anticheat.classifier;
 
 public interface Classifier {
 
-    int predict(double[] features);
+    int predict(long[] features);
 
 }
